test: Separate directory for test files
