--- .checkov.yaml
@@ -1,1 +1,0 @@
-skip-check:
