+++ tests/components/Message.test.ts
@@ -1,7 +1,7 @@
 // @vitest-environment jsdom
 import { render } from '@testing-library/svelte';
 import { expect, it } from 'vitest';
-import Message from './Message.svelte';
+import Message from '~/components/Message.svelte';
 
 it('renders OK', () => {
 	const { container } = render(Message);
