+++ tests/lib/chrome/bookmark.test.ts
@@ -1,5 +1,5 @@
 import { describe, it } from 'vitest';
-import { createBookmarks } from './bookmark';
+import { createBookmarks } from '../../../src/lib/chrome/bookmark';
 
 describe(createBookmarks, () => {
 	it.todo('creates bookmark recursively');
