+++ tests/lib/chrome/extension.test.ts
@@ -1,6 +1,10 @@
 import path from 'path';
 import { describe, expect, it } from 'vitest';
-import { getExtensionID, getExtensionPath, getProjectRoot } from './extension';
+import {
+	getExtensionID,
+	getExtensionPath,
+	getProjectRoot
+} from '../../../src/lib/chrome/extension';
 
 describe(getProjectRoot, () => {
 	it('should be', () => {
